Modified nav bar for login
Have modified the nav bar for when the person is logged in, have removed repetitive code but still doesn't fix my problem of removing profile link on profile page and same with accounts on accounts page. however I have moved all of the nav stuff on right side into a single file so there is less code repeated.